utilities: kernelscan.sh: Check that the file exists
kernelscan.sh makes assumptions about which files to scan and
this can change over time because files get renamed or removed from
the kernel. So add a check so see if the source exists before trying
to scan it.

Acked-by: Keng-Yu Lin <kengyu@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>

+++ src/utilities/kernelscan.sh
@@ -41,12 +41,16 @@ fi
 
 scan_source_file()
 {
-	$KERNELSCAN < $1 -E | gcc -E $CONFIGS - | $KERNELSCAN -P > $TMP
-	if [ $(stat -c%s $TMP) -gt 0 ]; then
-		echo "Source: $1"
-		cat $TMP
+	if [ -f $1 ]; then
+		$KERNELSCAN < $1 -E | gcc -E $CONFIGS - | $KERNELSCAN -P > $TMP
+		if [ $(stat -c%s $TMP) -gt 0 ]; then
+			echo "Source: $1"
+			cat $TMP
+		fi
+		rm $TMP	
+	else
+		echo "Source: $1 does not exist"
 	fi
-	rm $TMP
 }
 
 scan_source_tree()
